1316 1388 fix (#861)
* 1316/1388 Fix

Identified the cause of tickets 1316 and 1388. The issue was that, when a tube on the collection page was checked, scanned with the wrong tube ID, and then unchecked before the page was saved, the tube would be saved with the incorrect ID. The shipping manifest page checks that that value has been set and actually does not check that the tube is marked as collected. When user error caused a tube ID to be incorrectly scanned to one location, unchecked, and then correctly scanned to the next location, this was causing duplicate tube IDs to appear on the shipping manifest.

Code hardened in the following ways to prevent this:

shared.js:
* removeUnusableTubes updated to remove tubes which have not been marked as collected, as tubes which are not collected are in fact unusable.

events.js:
* collectionSubmission code updated so that the scanned ID will not be saved for tubes whose barcodes have errors.
* Logic further down also added so that if a tube is not checked but has a barcode, that barcode will be cleared.

Also updated some typos in shipping.js which are immaterial but were annoying me.

* Apply suggestions from code review

Spacing of if statements updated per convention.

Co-authored-by: Gene Barra <33293205+Gbarra9@users.noreply.github.com>

---------

Co-authored-by: Gene Barra <33293205+Gbarra9@users.noreply.github.com>